Issue #70 Split up Kafka consumers
